backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas>=2.2.0; python_version<"3.13"
reportlab==4.0.7
